requirements.txt
flask
kafka-python
pandas
scikit-learn
joblib
influxdb-client
